+++ src/Models/EntryAsset.php
@@ -47,7 +47,7 @@ class EntryAsset extends Asset
     }
 
     #[Override]
-    public function getTrailModelType(): string
+    public function getAdminType(): string
     {
         return Yii::t('media', 'ASSET_ENTRY_ASSET');
     }