+++ src/Models/SectionAsset.php
@@ -44,7 +44,7 @@ class SectionAsset extends Asset
     }
 
     #[Override]
-    public function getTrailModelType(): string
+    public function getAdminType(): string
     {
         return Yii::t('media', 'ASSET_SECTION_ASSET');
     }